requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
PyMySQL==1.1.0